Move files and delete the security package
