--- src/main/java/fr/dta/ovg/security/configurations/package-info.java
@@ -1,7 +1,0 @@
-/* Sub Package info Security Configurations.
- * @author Colin Cerveaux @C-ambium.
- * License : ©2019 All rights reserved.
- */
-
-/** This sub package contains security configuration classes.*/
-package fr.dta.ovg.security.configurations;
