--- src/main/java/fr/dta/ovg/security/package-info.java
@@ -1,7 +1,0 @@
-/* Package info Security.
- * @author Colin Cerveaux @C-ambium.
- * License : ©2019 All rights reserved.
- */
-
-/** This package contains all the security classes.*/
-package fr.dta.ovg.security;
